testing added + parser fixed and moved

+++ src/main/java/org/rogerthat/endpoints/FileEndpoint.java
@@ -6,6 +6,7 @@ import org.jboss.resteasy.plugins.providers.multipart.InputPart;
 import org.jboss.resteasy.plugins.providers.multipart.MultipartFormDataInput;
 import org.rogerthat.orm.CsvFiles;
 import org.rogerthat.orm.User;
+import org.rogerthat.services.CSVParser;
 
 import javax.transaction.Transactional;
 import javax.ws.rs.Consumes;
@@ -40,7 +41,7 @@ public class FileEndpoint {
 	@Path("/upload")
 	@Consumes("multipart/form-data")
 	@Transactional
-	public Response uploadFile(MultipartFormDataInput input, @QueryParam("userId") int userId) {
+	public Response uploadFile(MultipartFormDataInput input, @QueryParam("userId") Long userId) {
 
 		String fileName = "";
 		String csvFileName = "";
